Add files via upload
Took a raw coffee sales order data and created a dashboard to make it much more understandable and interactive for stakeholders or viewers that want to understand it more. With the dashboard, stakeholder or viewers can easily get an insight to sales made, top customers, sales by country, sales over time and many more.